+++ src/backend/marsha/bbb/tests/api/classroom/recordings/test_create_vod.py
@@ -1,8 +1,11 @@
 """Tests for the ClassroomRecording create vod API."""
 from unittest import mock
 
+from django.core.cache import cache
 from django.test import TestCase, override_settings
 from django.utils import timezone
+
+import responses
 
 from marsha.bbb.factories import ClassroomFactory, ClassroomRecordingFactory
 from marsha.core.defaults import PENDING
@@ -21,12 +24,20 @@ from marsha.core.utils.time_utils import to_timestamp
 # pylint: disable=unused-argument
 
 
+@override_settings(BBB_API_ENDPOINT="https://10.7.7.1/bigbluebutton/api")
+@override_settings(BBB_API_SECRET="SuperSecret")
+@override_settings(BBB_API_CALLBACK_SECRET="OtherSuperSecret")
+@override_settings(BBB_ENABLED=True)
 @override_settings(
     BBB_ENABLED=True,
     AWS_S3_REGION_NAME="us-east-1",
     AWS_SOURCE_BUCKET_NAME="test-source-bucket",
-    AWS_ACCESS_KEY_ID="test",
-    AWS_SECRET_ACCESS_KEY="test",
+    CACHES={
+        "default": {
+            "BACKEND": "django.core.cache.backends.locmem.LocMemCache",
+        },
+        "memory_cache": {"BACKEND": "django.core.cache.backends.locmem.LocMemCache"},
+    },
 )
 class ClassroomRecordingCreateVodAPITest(TestCase):
     """Test for the ClassroomRecording create vod API."""
@@ -39,6 +50,63 @@ class ClassroomRecordingCreateVodAPITest(TestCase):
 
         # Force URLs reload to use BBB_ENABLED
         reload_urlconf()
+
+        # Clear cache
+        cache.clear()
+
+        # An additional call is made since we added the url retrieval from the cache to the
+        # serializer
+        responses.add(
+            responses.GET,
+            "https://10.7.7.1/bigbluebutton/api/getRecordings",
+            match=[
+                responses.matchers.query_param_matcher(
+                    {
+                        "recordID": "67df5782-c17b-46d8-9dcb-a404e0b31251",
+                        "checksum": "5b9680a8fcca9e43f41f494b0503a41c14a86be9",
+                    }
+                )
+            ],
+            body="""
+            <response>
+                <returncode>SUCCESS</returncode>
+                <recordings>
+                    <recording>
+                        <recordID>c62c9c205d37815befe1b75ae6ef5878d8da5bb6-1673282694493</recordID>
+                        <meetingID>7e1c8b28-cd7a-4abe-93b2-3121366cb049</meetingID>
+                        <internalMeetingID>c62c9c205d37815befe1b75ae6ef5878d8da5bb6-1673282694493</internalMeetingID>
+                        <name>test ncl</name>
+                        <published>true</published>
+                        <state>published</state>
+                        <startTime>1673282694493</startTime>
+                        <endTime>1673282727208</endTime>
+                        <participants>1</participants>
+                        <metadata>
+                            <analytics-callback-url>https://10.7.7.2/bbb-analytics/api/v1/post_events?tag=bbb-dev
+                            </analytics-callback-url>
+                        </metadata>
+                        <playback>
+                            <format>
+                                <type>presentation</type>
+                                <url>
+                                    https://10.7.7.1/playback/presentation/2.3/c62c9c205d37815befe1b75ae6ef5878d8da5bb6-1673282694493
+                                </url>
+                                <length>0</length>
+                            </format>
+                            <format>
+                                <type>video</type>
+                                <url>
+                                    https://10.7.7.1/presentation/c62c9c205d37815befe1b75ae6ef5878d8da5bb6-1673282694493/meeting.mp4
+                                </url>
+                                <length>0</length>
+                            </format>
+                        </playback>
+                    </recording>
+                </recordings>
+            </response>
+            """,
+            status=200,
+        )
 
     def test_api_classroom_recording_create_anonymous(self):
         """An anonymous should not be able to convert a recording to a VOD."""
@@ -94,10 +162,12 @@ class ClassroomRecordingCreateVodAPITest(TestCase):
             },
         )
 
+    @responses.activate
     def test_api_classroom_recording_create_vod_instructor_or_admin(self):
         """Instructors and admins should be able to convert a recording to a VOD."""
         recording = ClassroomRecordingFactory(
             started_at="2019-08-21T15:00:02Z",
+            record_id="67df5782-c17b-46d8-9dcb-a404e0b31251",
         )
         jwt_token = InstructorOrAdminLtiTokenFactory(
             playlist=recording.classroom.playlist
@@ -131,7 +201,10 @@ class ClassroomRecordingCreateVodAPITest(TestCase):
                 "id": str(recording.id),
                 "record_id": str(recording.record_id),
                 "started_at": "2019-08-21T15:00:02Z",
-                "video_file_url": recording.video_file_url,
+                "video_file_url": (
+                    "https://10.7.7.1/presentation/"
+                    "c62c9c205d37815befe1b75ae6ef5878d8da5bb6-1673282694493/meeting.mp4"
+                ),
                 "vod": {
                     "id": str(recording.vod.id),
                     "title": "My title",
@@ -143,7 +216,10 @@ class ClassroomRecordingCreateVodAPITest(TestCase):
         self.assertEqual(Video.objects.first(), recording.vod)
         self.assertEqual(recording.vod.upload_state, PENDING)
         mock_invoke_lambda_convert.assert_called_once_with(
-            recording.video_file_url,
+            (
+                "https://10.7.7.1/presentation/"
+                "c62c9c205d37815befe1b75ae6ef5878d8da5bb6-1673282694493/meeting.mp4"
+            ),
             recording.vod.get_source_s3_key(stamp=to_timestamp(now)),
         )
 
@@ -196,13 +272,15 @@ class ClassroomRecordingCreateVodAPITest(TestCase):
 
         self.assertEqual(response.status_code, 403)
 
+    @responses.activate
     def test_api_classroom_recording_create_vod_user_access_token_organization_admin(
         self,
     ):
         """An organization administrator should be able to convert a recording to a VOD."""
         organization_access = OrganizationAccessFactory(role=ADMINISTRATOR)
         recording = ClassroomRecordingFactory(
-            classroom__playlist__organization=organization_access.organization
+            classroom__playlist__organization=organization_access.organization,
+            record_id="67df5782-c17b-46d8-9dcb-a404e0b31251",
         )
         jwt_token = UserAccessTokenFactory(user=organization_access.user)
 
@@ -214,6 +292,7 @@ class ClassroomRecordingCreateVodAPITest(TestCase):
 
         self.assertEqual(response.status_code, 201)
 
+    @responses.activate
     def test_api_classroom_recording_create_vod_from_standalone_site_no_consumer_site(
         self,
     ):
@@ -226,6 +305,7 @@ class ClassroomRecordingCreateVodAPITest(TestCase):
             classroom__playlist__organization=organization_access.organization,
             classroom__playlist__consumer_site=None,
             classroom__playlist__lti_id=None,
+            record_id="67df5782-c17b-46d8-9dcb-a404e0b31251",
         )
         jwt_token = UserAccessTokenFactory(user=organization_access.user)
 
@@ -263,13 +343,15 @@ class ClassroomRecordingCreateVodAPITest(TestCase):
 
         self.assertEqual(response.status_code, 405)
 
+    @responses.activate
     def test_api_classroom_recording_create_vod_user_access_token_playlist_admin(
         self,
     ):
         """A playlist administrator should be able to convert a recording to a VOD."""
         playlist_access = PlaylistAccessFactory(role=ADMINISTRATOR)
         recording = ClassroomRecordingFactory(
-            classroom__playlist=playlist_access.playlist
+            classroom__playlist=playlist_access.playlist,
+            record_id="67df5782-c17b-46d8-9dcb-a404e0b31251",
         )
         jwt_token = UserAccessTokenFactory(user=playlist_access.user)
 
@@ -281,13 +363,15 @@ class ClassroomRecordingCreateVodAPITest(TestCase):
 
         self.assertEqual(response.status_code, 201)
 
+    @responses.activate
     def test_api_classroom_recording_create_vod_user_access_token_playlist_instructor(
         self,
     ):
         """A playlist instructor should be able to convert a recording to a VOD."""
         playlist_access = PlaylistAccessFactory(role=INSTRUCTOR)
         recording = ClassroomRecordingFactory(
-            classroom__playlist=playlist_access.playlist
+            classroom__playlist=playlist_access.playlist,
+            record_id="67df5782-c17b-46d8-9dcb-a404e0b31251",
         )
         jwt_token = UserAccessTokenFactory(user=playlist_access.user)